Decided to go ahead and begin using Paper, as I'll be implementing ALL navigation using Paper Button
