+++ components/SkillsHome/SkillGrade.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { StyleSheet, View, Text, Image } from 'react-native';
 
-interface Props {
-    
+type Props = {
+
 }
 
 export default function SkillGrade() {
